Add relative time printing in home-loop.php and custom-loop.php. #96
For a post that was added 2 hours ago, it will say "hace 2 horas".
There is a TODO in that the relative time may be incorrect come
DST, as the offset from UTC for Chicago will change from -5 hours
to -6 hours. Also, it would be good to make it handle whatever
the server's current TZ is, so this could be used anywhere.